fix (provider/gateway): move batch support to a GatewayBatchLanguageModel subclass
